improvements(addressbook): clean code / refactoring
