+++ src/atomic.dex.qt.wallet.manager.hpp
@@ -22,6 +22,7 @@
 #include <QVariantMap>
 
 //! Project Headers
+#include "atomic.dex.mm2.hpp"
 #include "atomic.dex.security.hpp"
 #include "atomic.dex.version.hpp"
 #include "atomic.dex.wallet.config.hpp"
@@ -31,31 +32,38 @@ namespace atomic_dex
     class qt_wallet_manager
     {
       public:
+        QString get_wallet_default_name() const noexcept;
+
+        void set_wallet_default_name(QString wallet_name) noexcept;
+
         template <typename Functor>
-        bool inline login(
-            const QString& password, const QString& wallet_name, mm2& mm2_system, const QString& default_wallet_name, Functor&& login_if_success_functor);
+        bool
+        login(const QString& password, const QString& wallet_name, mm2& mm2_system, const QString& default_wallet_name, Functor&& login_if_success_functor);
 
-        inline bool load_wallet_cfg(const std::string& wallet_name);
+        bool create(const QString& password, const QString& seed, const QString& wallet_name);
 
-        inline QStringList get_categories_list() const noexcept;
+        bool load_wallet_cfg(const std::string& wallet_name);
 
-        inline QVariantMap get_address_from(const std::string& contact_name) const noexcept;
+        QStringList get_categories_list() const noexcept;
 
-        inline bool add_category(const std::string& category_name, bool with_update_file) noexcept;
+        QVariantMap get_address_from(const std::string& contact_name) const noexcept;
 
-        static inline QStringList get_wallets() noexcept;
+        bool add_category(const std::string& category_name, bool with_update_file) noexcept;
 
-        static inline bool is_there_a_default_wallet() noexcept;
+        static QStringList get_wallets() noexcept;
 
-        static inline QString get_default_wallet_name() noexcept;
+        static bool is_there_a_default_wallet() noexcept;
 
-        static inline bool delete_wallet(const QString& wallet_name) noexcept;
+        static QString get_default_wallet_name() noexcept;
 
-        static inline bool confirm_password(const QString& wallet_name, const QString& password);
+        static bool delete_wallet(const QString& wallet_name) noexcept;
+
+        static bool confirm_password(const QString& wallet_name, const QString& password);
 
       private:
-        inline bool update_wallet_cfg() noexcept;
-        wallet_cfg  m_wallet_cfg;
+        bool       update_wallet_cfg() noexcept;
+        wallet_cfg m_wallet_cfg;
+        QString    m_current_default_wallet{""};
     };
 
     template <typename Functor>
@@ -100,144 +108,6 @@ namespace atomic_dex
             load_wallet_cfg(default_wallet.toStdString());
             mm2_system.spawn_mm2_instance(default_wallet.toStdString(), seed);
             return true;
-        }
-        return false;
-    }
-
-    QStringList
-    qt_wallet_manager::get_wallets() noexcept
-    {
-        QStringList out;
-
-        for (auto&& p: fs::directory_iterator((get_atomic_dex_config_folder())))
-        {
-            if (p.path().extension().string() == ".seed")
-            {
-                out.push_back(QString::fromStdString(p.path().stem().string()));
-            }
-        }
-
-        return out;
-    }
-
-    bool
-    qt_wallet_manager::is_there_a_default_wallet() noexcept
-    {
-        return fs::exists(get_atomic_dex_config_folder() / "default.wallet");
-    }
-
-    QString
-    qt_wallet_manager::get_default_wallet_name() noexcept
-    {
-        if (is_there_a_default_wallet())
-        {
-            std::ifstream ifs((get_atomic_dex_config_folder() / "default.wallet").c_str());
-            assert(ifs);
-            std::string str((std::istreambuf_iterator<char>(ifs)), std::istreambuf_iterator<char>());
-            return QString::fromStdString(str);
-        }
-        return "nonexistent";
-    }
-
-    bool
-    qt_wallet_manager::delete_wallet(const QString& wallet_name) noexcept
-    {
-        using namespace std::string_literals;
-        return fs::remove(get_atomic_dex_config_folder() / (wallet_name.toStdString() + ".seed"s));
-    }
-
-    bool
-    qt_wallet_manager::confirm_password(const QString& wallet_name, const QString& password)
-    {
-        std::error_code ec;
-        auto            key = atomic_dex::derive_password(password.toStdString(), ec);
-        if (ec)
-        {
-            spdlog::debug("{}", ec.message());
-            if (ec == dextop_error::derive_password_failed)
-            {
-                return false;
-            }
-        }
-        using namespace std::string_literals;
-        const fs::path seed_path = get_atomic_dex_config_folder() / (wallet_name.toStdString() + ".seed"s);
-        auto           seed      = atomic_dex::decrypt(seed_path, key.data(), ec);
-        if (ec == dextop_error::corrupted_file_or_wrong_password)
-        {
-            spdlog::warn("{}", ec.message());
-            return false;
-        }
-        return true;
-    }
-
-    bool
-    qt_wallet_manager::load_wallet_cfg(const std::string& wallet_name)
-    {
-        using namespace std::string_literals;
-        const fs::path wallet_object_path = get_atomic_dex_export_folder() / (wallet_name + ".wallet.json"s);
-        std::ifstream  ifs(wallet_object_path.string());
-
-        if (not ifs.is_open())
-        {
-            return false;
-        }
-        nlohmann::json j;
-        ifs >> j;
-        m_wallet_cfg = j;
-        return true;
-    }
-
-    QStringList
-    qt_wallet_manager::get_categories_list() const noexcept
-    {
-        QStringList out;
-
-        out.reserve(m_wallet_cfg.categories_addressbook_registry.size());
-        for (const auto& [key, _]: m_wallet_cfg.categories_addressbook_registry)
-        {
-            (void)_;
-            out.push_back(QString::fromStdString(key));
-        }
-        return out;
-    }
-
-    QVariantMap
-    qt_wallet_manager::get_address_from(const std::string& contact_name) const noexcept
-    {
-        QVariantMap out;
-
-        if (m_wallet_cfg.addressbook_registry.find(contact_name) != m_wallet_cfg.addressbook_registry.cend())
-        {
-            for (auto&& [key, value]: m_wallet_cfg.addressbook_registry.at(contact_name))
-            { out.insert(QString::fromStdString(key), QVariant(QString::fromStdString(value))); }
-        }
-        return out;
-    }
-
-    bool
-    qt_wallet_manager::update_wallet_cfg() noexcept
-    {
-        using namespace std::string_literals;
-        const fs::path wallet_object_path = get_atomic_dex_export_folder() / (m_wallet_cfg.name + ".wallet.json"s);
-        std::ofstream  ofs(wallet_object_path.string(), std::ios::trunc);
-        if (not ofs.is_open())
-        {
-            return false;
-        }
-
-        nlohmann::json j;
-        atomic_dex::to_json(j, m_wallet_cfg);
-        ofs << j.dump(4);
-        return true;
-    }
-
-    bool
-    qt_wallet_manager::add_category(const std::string& category_name, bool with_update_file) noexcept
-    {
-        if (m_wallet_cfg.categories_addressbook_registry.find(category_name) == m_wallet_cfg.categories_addressbook_registry.cend())
-        {
-            m_wallet_cfg.categories_addressbook_registry[category_name] = {};
-            return with_update_file ? update_wallet_cfg() : true;
         }
         return false;
     }
